+++ src/bigquery_cleaner/__init__.py
@@ -4,7 +4,7 @@ import logging
 
 __all__ = ["__version__", "logger"]
 
-__version__ = "0.1.8"
+__version__ = "0.1.9"
 
 # Configure logging
 logger = logging.getLogger("bigquery_cleaner")